requirements.txt
langchain==0.1.12
langchain-community==0.0.27
langchain-groq==0.0.1
langchain-huggingface==0.0.2
crewai==0.28.0
crewai-tools==0.0.5
faiss-cpu==1.7.4
python-dotenv==1.0.1
pypdf==4.0.2
sentence-transformers==2.5.1 